refactor(handlers): extract court and navigation handlers (WP5.9.5.4)
Extract court-related and navigation handlers from useRegistrationHandlers.js
into separate modules, completing the handler extraction phase.

Court handlers:
- saveCourtData, getAvailableCourts, clearViaService (internal)
- clearCourt, assignCourtToGroup, changeCourt, sendGroupToWaitlist

Navigation handlers:
- checkLocationAndProceed, handleToggleAddPlayer, handleGroupGoBack

useRegistrationHandlers.js: 751 → 479 lines (meets ≤499 stop rule)
