+++ src/registration/appHandlers/handlers/index.js
@@ -1,3 +1,5 @@
 export { useAdminHandlers } from './adminHandlers.js';
 export { useGuestHandlers } from './guestHandlers.js';
 export { useGroupHandlers } from './groupHandlers.js';
+export { useCourtHandlers } from './courtHandlers.js';
+export { useNavigationHandlers } from './navigationHandlers.js';
